Experimental changes that use symbols for attribute and element names.
It is not clear how best to handle colons in attribute and element names.
Colons (:) are allowed as namespace separators, but in Nu, colons may
only be the last character of a symbol.  In these changes, there is code
to replace colons with some other character, but that is currently disabled.

There are also two new examples, one showing the use of namespaces
and one that uses NuSAX to read the result of a call to the github api.